# docs/openapi.yaml
openapi: "3.0.3"
info:
  title: KnowFlow — 智能学习助手平台 API
  description: |
    KnowFlow 是一个基于 Spring AI 2.0 与 RAG（检索增强生成）技术的全栈智能学习助手平台。
    本 API 文档定义了 KnowFlow 后端所有 RESTful 接口规范。

    ### 核心功能
    - **用户管理**：注册、登录、个人信息获取
    - **智能对话**：多轮连续对话、SSE 流式输出、对话历史管理
    - **RAG 知识库**：文档上传解析、基于知识库的精准问答、文档管理

    ### 通用约定
    | 规范项 | 约定 |
    |:---|:---|
    | 基础路径 | `/api` |
    | 请求体格式 | `application/json`（文件上传除外） |
    | 响应格式 | 统一 `ApiResponse<T>` 结构 |
    | 鉴权方式 | Bearer Token（JWT） |
    | 流式传输 | `text/event-stream`（SSE） |
  version: "1.0.0"
  contact:
    name: KnowFlow Team
    url: https://github.com/cloudtheon/knowflow
servers:
  - url: http://localhost:8080/api
    description: 本地开发环境
  - url: https://api.knowflow.dev/api
    description: 生产环境

# =====================================================================
# 全局组件
# =====================================================================
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 登录成功后获取 JWT Token，以 `Bearer <token>` 格式放入请求头
  schemas:
...
    ApiResponse:
      type: object
      description: 统一 API 响应包装
      properties:
        code:
          type: integer
          description: 业务状态码（200 成功，400 参数错误，401 未授权，403 无权限，404 资源不存在，409 资源冲突，500 服务器错误）
          example: 200
        message:
          type: string
          description: 响应消息
          example: "success"
        data:
          type: object
          description: 响应数据（具体结构见各接口定义）
      required:
        - code
        - message
        - data

    # ---------- 分页响应 ----------
...
    UserInfo:
      type: object
      description: 用户基本信息
      properties:
        id:
          type: integer
          format: int64
          description: 用户 ID
          example: 1
        username:
          type: string
          description: 用户名
          example: "zhangsan"
        avatar:
          type: string
          description: 头像 URL
          example: "https://example.com/avatar.png"
      required:
        - id
        - username
    LoginRequest:
      type: object
      description: 登录请求
      properties:
        username:
          type: string
          description: 用户名
          example: "zhangsan"
        password:
          type: string
          description: 密码
          example: "password123"
      required:
        - username
        - password
    RegisterRequest:
      type: object
      description: 注册请求
      properties:
        username:
          type: string
          description: 用户名（3~50 字符，字母/数字/下划线）
          example: "zhangsan"
        password:
          type: string
          description: 密码（6~100 字符）
          example: "password123"
      required:
        - username
        - password
    LoginResponse:
      type: object
      description: 登录/注册响应数据
      properties:
        token:
          type: string
          description: JWT Token
          example: "eyJhbGciOiJIUzI1NiIs..."
        user:
          $ref: "#/components/schemas/UserInfo"
      required:
        - token
        - user

    # ---------- 对话相关 ----------
...
    ErrorResponse:
      type: object
      description: 错误响应
      properties:
        code:
          type: integer
          description: 业务状态码
          example: 400
        message:
          type: string
          description: 错误描述
          example: "用户名已存在"
        data:
          type: object
          nullable: true
          description: 错误详情（可选）
          example: null
      required:
        - code
        - message
        - data

# =====================================================================
# 路径定义
# =====================================================================
paths:
...
  /auth/register:
    post:
      tags:
        - 01-用户认证
      summary: 用户注册
      description: "创建新用户。注册成功后自动返回 JWT Token，前端可跳转至聊天页。\n\n**注意事项**：\n- 用户名必须唯一\n- 密码需经过 BCrypt 加密后存储"
      operationId: registerUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
            example:
              username: "zhangsan"
              password: "password123"
      responses:
        "200":
          description: 注册成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/LoginResponse"
              example:
                code: 200
                message: "success"
                data:
                  token: "eyJhbGciOiJIUzI1NiIs..."
                  user:
                    id: 1
                    username: "zhangsan"
                    avatar: null
        "409":
          description: 用户名已存在
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                code: 409
                message: "用户名已存在"
                data: null
  /auth/login:
    post:
      tags:
        - 01-用户认证
      summary: 用户登录
      description: 使用用户名和密码登录，返回 JWT Token 和用户信息。Token 有效期默认为 7 天。
      operationId: loginUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
            example:
              username: "zhangsan"
              password: "password123"
      responses:
        "200":
          description: 登录成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/LoginResponse"
              example:
                code: 200
                message: "success"
                data:
                  token: "eyJhbGciOiJIUzI1NiIs..."
                  user:
                    id: 1
                    username: "zhangsan"
                    avatar: null
        "401":
          description: 用户名或密码错误
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                code: 401
                message: "用户名或密码错误"
                data: null
  /auth/profile:
    get:
      tags:
        - 01-用户认证
      summary: 获取当前用户信息
      description: 获取当前登录用户的基本信息。需携带有效的 JWT Token。
      operationId: getUserProfile
      security:
        - BearerAuth: []
      responses:
        "200":
          description: 获取成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/UserInfo"
              example:
                code: 200
                message: "success"
                data:
                  id: 1
                  username: "zhangsan"
                  avatar: null
        "401":
          description: Token 无效或已过期
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                code: 401
                message: "未授权，请重新登录"
                data: null

  # ==================== 对话模块 ====================
...