# knowflow-backend/knowflow-core/src/main/java/com/cloudtheon/knowflowcore/dto/RegisterRequest.java
package com.cloudtheon.knowflowcore.dto;

import io.swagger.v3.oas.annotations.media.Schema;
import jakarta.validation.constraints.NotBlank;
import jakarta.validation.constraints.Size;
import lombok.Data;

/**
 * 注册请求 DTO
 */
@Data
@Schema(description = "注册请求")
public class RegisterRequest {

    @NotBlank(message = "用户名不能为空")
    @Size(min = 3, max = 50, message = "用户名长度需在 3~50 个字符之间")
    @Schema(description = "用户名（3~50 字符，字母/数字/下划线）", example = "zhangsan")
    private String username;

    @NotBlank(message = "密码不能为空")
    @Size(min = 6, max = 100, message = "密码长度需在 6~100 个字符之间")
    @Schema(description = "密码（6~100 字符）", example = "password123")
    private String password;
}